+++ glade_forms/lib/src/core/changes_info.dart
@@ -2,6 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:glade_forms/src/validator/validator_result.dart';
 
 class ChangesInfo<T> extends Equatable {
+  final String inputKey;
   final T? initialValue;
   final T? previousValue;
 
@@ -10,8 +11,10 @@ class ChangesInfo<T> extends Equatable {
   final ValidatorResult<T>? validatorResult;
 
   @override
-  List<Object?> get props => [initialValue, previousValue, value, validatorResult];
+  List<Object?> get props => [initialValue, previousValue, value, validatorResult, inputKey];
+
   const ChangesInfo({
+    required this.inputKey,
     required this.previousValue,
     required this.value,
     required this.validatorResult,