+++ storybook/lib/main.dart
@@ -5,10 +5,11 @@ import 'package:flutter/material.dart';
 import 'package:glade_forms_storybook/generated/locale_loader.g.dart';
 import 'package:glade_forms_storybook/localization_addon_custom.dart';
 import 'package:glade_forms_storybook/usecases/complex_object_mapping_example.dart';
-import 'package:glade_forms_storybook/usecases/one_checkbox_deps_validation.dart';
+import 'package:glade_forms_storybook/usecases/dependencies/checkbox_dependency_change.dart';
+import 'package:glade_forms_storybook/usecases/onchange/one_checkbox_deps_validation.dart';
+import 'package:glade_forms_storybook/usecases/onchange/two_way_checkbox_change.dart';
 import 'package:glade_forms_storybook/usecases/quickstart_example.dart';
 import 'package:glade_forms_storybook/usecases/regress/issue48_text_controller_example.dart';
-import 'package:glade_forms_storybook/usecases/two_way_checkbox_change.dart';
 import 'package:widgetbook/widgetbook.dart';
 
 // ignore: prefer-static-class, ok for now
@@ -62,15 +63,24 @@ class App extends StatelessWidget {
       directories: [
         WidgetbookUseCase(name: 'Quickstart form', builder: (context) => const QuickStartExample()),
         WidgetbookCategory(
-          name: 'Dependencies',
+          name: 'onChange',
           children: [
             WidgetbookUseCase(
-              name: 'One checkbox dependency',
+              name: 'One checkbox onChange',
               builder: (context) => const OneCheckboxValidationDependencyExample(),
             ),
             WidgetbookUseCase(
-              name: 'Two-way checkbox dependency',
+              name: 'Two-way checkbox onChange',
               builder: (context) => const TwoWayCheckboxExample(),
+            ),
+          ],
+        ),
+        WidgetbookCategory(
+          name: 'Dependency',
+          children: [
+            WidgetbookUseCase(
+              name: 'Checkbox dependency',
+              builder: (context) => const CheckboxDependencyExample(),
             ),
           ],
         ),